added a behaviour in the today screen so that the weatherforecast is directly shown starting from the current hour, shrinked temperature values in the home screen to show just the int

+++ app/src/main/java/co/develhope/meteoapp/features/home/HomeScreen.kt
@@ -52,8 +52,8 @@ class HomeScreen : Fragment() {
     }
 
     fun showTodaysMeteo(item: WeatherConditions, sharedPreferencesHelper:SharedPreferencesHelper,position: Int = 0) {
-        val maxTemperature = item.daily.maxTemperature[position]
-        val minTemperature = item.daily.minTemperature[position]
+        val maxTemperature = item.daily.maxTemperature[position].toInt()
+        val minTemperature = item.daily.minTemperature[position].toInt()
         val date = item.daily.time[position]
         val rain = item.daily.rainSum[position]
         val wind = item.daily.windMax[position]
